Catapult funcitonality completed
Catapult functionality has been completed. A launchable object can be dragged into the catapult arm. The launchable is then launched when the user interacts with the catapult. The developer can position the aim in the level editor for the catapult to be shot at. Also edited the DraggedObject base class so that the velocity limits can be broken if the programmer has to.